Refactor some GUI/button stuff
- ButtonBase doesn't extend the vanilla GuiButton anymore, but instead
  it now extends WidgetBase
- The button listener interface isn't generic anymore
- The buttons are now mostly handled as widgets

+++ src/main/java/fi/dy/masa/malilib/config/gui/ConfigPanelBase.java
@@ -326,7 +326,7 @@ public abstract class ConfigPanelBase extends AbstractConfigPanel
         }
 
         @Override
-        public boolean onMouseScrolled(int mouseX, int mouseY, int mouseWheelDelta)
+        public boolean onMouseScrolled(int mouseX, int mouseY, double mouseWheelDelta)
         {
             return this.foregroundGui.onMouseScrolled(mouseX, mouseY, mouseWheelDelta);
         }
